update & fix existing tests
